chore: Move register snap endpoint to endpoints section of app
