+++ tests/endpoints/publisher/tests_register.py
