+++ webapp/publisher/snaps/views.py
@@ -34,6 +34,7 @@ from webapp.endpoints.publisher.settings import (
 )
 from webapp.endpoints.publisher.publicise import get_publicise_data
 from webapp.endpoints.publisher.packages import get_package_metadata
+from webapp.endpoints.publisher.register import post_register_name
 from webapp.endpoints import releases, builds
 from webapp.publisher.snaps.builds import map_snap_build_status
 
@@ -383,71 +384,16 @@ def redirect_post_register_name():
     return flask.redirect(flask.url_for(".post_register_name"), 307)
 
 
-@publisher_snaps.route("/api/register-snap", methods=["POST"])
-@login_required
-def post_register_name():
-    snap_name = flask.request.form.get("snap_name")
-    res = {}
-
-    if not snap_name:
-        res["success"] = False
-        res["message"] = "You must define a snap name"
-
-        return jsonify(res)
-
-    is_private = flask.request.form.get("is_private") == "private"
-    store = flask.request.form.get("store")
-
-    try:
-        dashboard.post_register_name(
-            session=flask.session,
-            snap_name=snap_name,
-            registrant_comment=None,
-            is_private=is_private,
-            store=store,
-        )
-    except StoreApiResponseErrorList as api_response_error_list:
-        res = {
-            "success": False,
-            "data": {
-                "is_private": is_private,
-                "snap_name": snap_name,
-                "store": store,
-            },
-        }
-
-        if api_response_error_list.status_code == 409:
-            for error in api_response_error_list.errors:
-                res["data"]["error_code"] = error["code"]
-
-                return jsonify(res)
-
-        if api_response_error_list.status_code == 400:
-            res["data"]["error_code"] = "no-permission"
-            res[
-                "message"
-            ] = """You don't have permission
-                to register a snap in this store.
-                Please see store administrator."""
-
-            return jsonify(res)
-
-        res["message"] = "Unable to register snap name"
-        res["data"] = {
-            "snap_name": snap_name,
-            "is_private": is_private,
-            "store": store,
-        }
-
-        return jsonify(res)
-
-    return jsonify({"success": True})
-
-
 publisher_snaps.add_url_rule(
     "/api/packages/<snap_name>",
     view_func=get_package_metadata,
     methods=["GET"],
+)
+
+publisher_snaps.add_url_rule(
+    "/api/register-snap",
+    view_func=post_register_name,
+    methods=["POST"],
 )
